Factor out HasOut common trait (#272)
