+++ features/src/main/scala/com/salesforce/op/stages/OpPipelineStages.scala
@@ -222,11 +222,10 @@ private[op] trait AllowLabelAsInput[O <: FeatureType] extends OpPipelineStage[O]
  * @tparam I input feature type
  * @tparam O output feature type
  */
-trait OpPipelineStage1[I <: FeatureType, O <: FeatureType] extends OpPipelineStage[O] with HasIn1 {
-  self: PipelineStage =>
+trait OpPipelineStage1[I <: FeatureType, O <: FeatureType]
+  extends OpPipelineStage[O] with HasIn1 with HasOut[O] {
 
-  implicit val tto: TypeTag[O]
-  implicit val ttov: TypeTag[O#Value]
+  self: PipelineStage =>
 
   final override type InputFeatures = FeatureLike[I]
 
@@ -236,16 +235,7 @@ trait OpPipelineStage1[I <: FeatureType, O <: FeatureType] extends OpPipelineSta
 
   protected[op] override def outputFeatureUid: String = FeatureUID[O](uid)
 
-  override def getOutput(): FeatureLike[O] = new Feature[O](
-    uid = outputFeatureUid,
-    name = getOutputFeatureName,
-    originStage = this,
-    isResponse = outputIsResponse,
-    parents = getInputFeatures()
-  )(tto)
-
 }
-
 
 /**
  * Pipeline stage of Feature type I to Features O1 and O2
@@ -319,11 +309,8 @@ trait OpPipelineStage1to3[I <: FeatureType, O1 <: FeatureType, O2 <: FeatureType
  * @tparam O  output feature type
  */
 trait OpPipelineStage2[I1 <: FeatureType, I2 <: FeatureType, O <: FeatureType]
-  extends OpPipelineStage[O] with HasIn1 with HasIn2 {
+  extends OpPipelineStage[O] with HasIn1 with HasIn2 with HasOut[O] {
   self: PipelineStage =>
-
-  implicit val tto: TypeTag[O]
-  implicit val ttov: TypeTag[O#Value]
 
   final override type InputFeatures = (FeatureLike[I1], FeatureLike[I2])
 
@@ -334,14 +321,6 @@ trait OpPipelineStage2[I1 <: FeatureType, I2 <: FeatureType, O <: FeatureType]
   }
 
   protected[op] override def outputFeatureUid: String = FeatureUID[O](uid)
-
-  override def getOutput(): FeatureLike[O] = new Feature[O](
-    uid = outputFeatureUid,
-    name = getOutputFeatureName,
-    originStage = this,
-    isResponse = outputIsResponse,
-    parents = getInputFeatures()
-  )(tto)
 
 }
 
@@ -412,6 +391,7 @@ trait OpPipelineStage2to3[I1 <: FeatureType, I2 <: FeatureType, O1 <: FeatureTyp
   }
 }
 
+
 /**
  * Pipeline stage of Feature type I1, I2 and I3 to Feature of type O
  *
@@ -421,11 +401,8 @@ trait OpPipelineStage2to3[I1 <: FeatureType, I2 <: FeatureType, O1 <: FeatureTyp
  * @tparam O  output feature type
  */
 trait OpPipelineStage3[I1 <: FeatureType, I2 <: FeatureType, I3 <: FeatureType, O <: FeatureType]
-  extends OpPipelineStage[O] with HasIn1 with HasIn2 with HasIn3 {
+  extends OpPipelineStage[O] with HasIn1 with HasIn2 with HasIn3 with HasOut[O] {
   self: PipelineStage =>
-
-  implicit val tto: TypeTag[O]
-  implicit val ttov: TypeTag[O#Value]
 
   final override type InputFeatures = (FeatureLike[I1], FeatureLike[I2], FeatureLike[I3])
 
@@ -437,13 +414,6 @@ trait OpPipelineStage3[I1 <: FeatureType, I2 <: FeatureType, I3 <: FeatureType, 
 
   protected[op] override def outputFeatureUid: String = FeatureUID[O](uid)
 
-  override def getOutput(): FeatureLike[O] = new Feature[O](
-    uid = outputFeatureUid,
-    name = getOutputFeatureName,
-    originStage = this,
-    isResponse = outputIsResponse,
-    parents = getInputFeatures()
-  )(tto)
 }
 
 /**
@@ -490,11 +460,8 @@ trait OpPipelineStage3to2[I1 <: FeatureType, I2 <: FeatureType, I3 <: FeatureTyp
  * @tparam O  output feature type
  */
 trait OpPipelineStage4[I1 <: FeatureType, I2 <: FeatureType, I3 <: FeatureType, I4 <: FeatureType, O <: FeatureType]
-  extends OpPipelineStage[O] with HasIn1 with HasIn2 with HasIn3 with HasIn4 {
+  extends OpPipelineStage[O] with HasIn1 with HasIn2 with HasIn3 with HasIn4 with HasOut[O] {
   self: PipelineStage =>
-
-  implicit val tto: TypeTag[O]
-  implicit val ttov: TypeTag[O#Value]
 
   final override type InputFeatures = (FeatureLike[I1], FeatureLike[I2], FeatureLike[I3], FeatureLike[I4])
 
@@ -506,13 +473,6 @@ trait OpPipelineStage4[I1 <: FeatureType, I2 <: FeatureType, I3 <: FeatureType, 
 
   protected[op] override def outputFeatureUid: String = FeatureUID[O](uid)
 
-  override def getOutput(): FeatureLike[O] = new Feature[O](
-    uid = outputFeatureUid,
-    name = getOutputFeatureName,
-    originStage = this,
-    isResponse = outputIsResponse,
-    parents = getInputFeatures()
-  )(tto)
 }
 
 
@@ -522,11 +482,9 @@ trait OpPipelineStage4[I1 <: FeatureType, I2 <: FeatureType, I3 <: FeatureType, 
  * @tparam I input feature type
  * @tparam O output feature type
  */
-trait OpPipelineStageN[I <: FeatureType, O <: FeatureType] extends OpPipelineStage[O] with HasInN {
+trait OpPipelineStageN[I <: FeatureType, O <: FeatureType]
+  extends OpPipelineStage[O] with HasInN with HasOut[O] {
   self: PipelineStage =>
-
-  implicit val tto: TypeTag[O]
-  implicit val ttov: TypeTag[O#Value]
 
   final override type InputFeatures = Array[FeatureLike[I]]
 
@@ -540,13 +498,6 @@ trait OpPipelineStageN[I <: FeatureType, O <: FeatureType] extends OpPipelineSta
 
   protected[op] override def outputFeatureUid: String = FeatureUID[O](uid)
 
-  override def getOutput(): FeatureLike[O] = new Feature[O](
-    uid = outputFeatureUid,
-    name = getOutputFeatureName,
-    originStage = this,
-    isResponse = outputIsResponse,
-    parents = getInputFeatures()
-  )(tto)
 }
 
 /**
@@ -556,12 +507,9 @@ trait OpPipelineStageN[I <: FeatureType, O <: FeatureType] extends OpPipelineSta
  * @tparam I2 input sequence feature type
  * @tparam O output feature type
  */
-trait OpPipelineStage2N[I1 <: FeatureType, I2 <: FeatureType, O <: FeatureType] extends OpPipelineStage[O]
-  with HasIn1PlusN {
+trait OpPipelineStage2N[I1 <: FeatureType, I2 <: FeatureType, O <: FeatureType]
+  extends OpPipelineStage[O] with HasIn1PlusN with HasOut[O] {
   self: PipelineStage =>
-
-  implicit val tto: TypeTag[O]
-  implicit val ttov: TypeTag[O#Value]
 
   final override type InputFeatures = (FeatureLike[I1], Array[FeatureLike[I2]])
 
@@ -576,13 +524,6 @@ trait OpPipelineStage2N[I1 <: FeatureType, I2 <: FeatureType, O <: FeatureType] 
 
   protected[op] override def outputFeatureUid: String = FeatureUID[O](uid)
 
-  override def getOutput(): FeatureLike[O] = new Feature[O](
-    uid = outputFeatureUid,
-    name = getOutputFeatureName,
-    originStage = this,
-    isResponse = outputIsResponse,
-    parents = getInputFeatures()
-  )(tto)
 }
